requirements.txt
amqp==1.4.9
anyjson==0.3.3
beacontools==1.3.1
billiard==3.3.0.23
celery==3.1.26.post2
construct==2.9.45
Django==2.2.1
djangorestframework==3.9.4
kombu==3.0.37
pkg-resources==0.0.0
PyBluez==0.22
pytz==2019.1
sqlparse==0.3.0
